requirements.txt
Django>=5.0
psycopg2-binary
celery
redis
playwright
openai
reportlab
django-tailwind
python-dotenv
requests
matplotlib
numpy

# Production dependencies
gunicorn
dj-database-url
whitenoise
paramiko
